backend/code/requirements.txt
flask
flask-cors
argon2-cffi
flask-jwt-extended
python-dotenv